remove redundant tests in favor of parameterized tests
