+++ tests/integration/alb/test_alb_update.py
@@ -22,171 +22,6 @@ from tests.lib.update import (
 )
 
 
-@pytest.fixture
-def service_instance():
-    service_instance = factories.ALBServiceInstanceFactory.create(
-        id="4321",
-        domain_names=["example.com", "foo.com"],
-        route53_alias_hosted_zone="Z2FDTNDATAQYW2",
-        alb_arn="alb-arn-1",
-        alb_listener_arn="alb-listener-arn-1",
-        domain_internal="fake1234.cloud.test",
-    )
-    current_cert = factories.CertificateFactory.create(
-        service_instance=service_instance,
-        private_key_pem="SOMEPRIVATEKEY",
-        iam_server_certificate_id="certificate_id",
-        iam_server_certificate_arn="certificate_arn",
-        iam_server_certificate_name="certificate_name",
-        leaf_pem="SOMECERTPEM",
-        fullchain_pem="FULLCHAINOFSOMECERTPEM",
-        id=1001,
-    )
-    factories.ChallengeFactory.create(
-        domain="example.com",
-        validation_contents="example txt",
-        certificate_id=1001,
-        answered=True,
-    )
-    factories.ChallengeFactory.create(
-        domain="foo.com",
-        validation_contents="foo txt",
-        certificate_id=1001,
-        answered=True,
-    )
-    service_instance.current_certificate = current_cert
-    db.session.add(service_instance)
-    db.session.add(current_cert)
-    db.session.commit()
-    db.session.expunge_all()
-    return service_instance
-
-
-def test_refuses_to_update_synchronously(client):
-    client.update_alb_instance("4321", accepts_incomplete="false")
-
-    assert "AsyncRequired" in client.response.body
-    assert client.response.status_code == 422
-
-
-def test_refuses_to_update_synchronously_by_default(client):
-    client.update_alb_instance("4321", accepts_incomplete="")
-
-    assert "AsyncRequired" in client.response.body
-    assert client.response.status_code == 422
-
-
-def test_refuses_to_update_with_duplicate_domains(client, dns, service_instance):
-    factories.ALBServiceInstanceFactory.create(domain_names=["foo.com", "bar.com"])
-    dns.add_cname("_acme-challenge.example.com")
-    dns.add_cname("_acme-challenge.foo.com")
-
-    client.update_alb_instance("4321", params={"domains": "example.com, foo.com"})
-
-    assert "already exists" in client.response.body, client.response.body
-    assert client.response.status_code == 400, client.response.body
-
-
-def test_duplicate_domain_check_ignores_self(client, dns, service_instance):
-    dns.add_cname("_acme-challenge.example.com")
-    dns.add_cname("_acme-challenge.foo.com")
-
-    client.update_alb_instance("4321", params={"domains": "example.com, foo.com"})
-
-    assert client.response.status_code == 200, client.response.body
-
-
-def test_duplicate_domain_check_ignores_deactivated(client, dns, service_instance):
-    factories.ALBServiceInstanceFactory.create(
-        domain_names="foo.com", deactivated_at=datetime.utcnow()
-    )
-    dns.add_cname("_acme-challenge.foo.com")
-
-    client.update_alb_instance("4321", params={"domains": "foo.com"})
-
-    assert client.response.status_code == 202, client.response.body
-
-
-def test_refuses_to_update_without_any_acme_challenge_CNAMEs(client, service_instance):
-    client.update_alb_instance("4321", params={"domains": "bar.com,foo.com"})
-
-    desc = client.response.json.get("description")
-    assert "CNAME" in desc
-    assert "_acme-challenge.foo.com " in desc
-    assert "_acme-challenge.foo.com.domains.cloud.test" in desc
-    assert "_acme-challenge.bar.com " in desc
-    assert "_acme-challenge.bar.com.domains.cloud.test" in desc
-    assert client.response.status_code == 400
-
-
-def test_refuses_to_update_without_one_acme_challenge_CNAME(
-    client, dns, service_instance
-):
-    dns.add_cname("_acme-challenge.foo.com")
-    client.update_alb_instance("4321", params={"domains": "bar.com,foo.com"})
-
-    desc = client.response.json.get("description")
-    assert "CNAME" in desc
-    assert "_acme-challenge.foo.com" not in desc
-    assert "_acme-challenge.bar.com " in desc
-    assert "_acme-challenge.bar.com.domains.cloud.test" in desc
-    assert client.response.status_code == 400
-
-
-def test_refuses_to_update_with_incorrect_acme_challenge_CNAME(
-    client, dns, service_instance
-):
-    dns.add_cname("_acme-challenge.bar.com")
-    dns.add_cname("_acme-challenge.foo.com", target="INCORRECT")
-    client.update_alb_instance("4321", params={"domains": "bar.com,foo.com"})
-
-    desc = client.response.json.get("description")
-
-    assert "is set incorrectly" in desc
-
-    assert " _acme-challenge.foo.com " in desc
-    assert " _acme-challenge.foo.com.domains.cloud.test" in desc
-    assert " INCORRECT" in desc
-
-    assert " _acme-challenge.bar.com" not in desc
-    assert client.response.status_code == 400
-
-
-def test_refuses_update_for_canceled_instance(client, dns, clean_db, service_instance):
-    dns.add_cname("_acme-challenge.bar.com")
-    dns.add_cname("_acme-challenge.foo.com")
-    service_instance.deactivated_at = datetime.now()
-    db.session.add(service_instance)
-    db.session.commit()
-
-    client.update_alb_instance("4321", params={"domains": "bar.com,foo.com"})
-
-    desc = client.response.json.get("description")
-    assert "canceled" in desc
-    assert client.response.status_code == 400
-
-
-def test_refuses_update_for_nonexistent_instance(client, dns):
-    dns.add_cname("_acme-challenge.bar.com")
-    dns.add_cname("_acme-challenge.foo.com")
-    client.update_alb_instance("4321", params={"domains": "bar.com,foo.com"})
-    desc = client.response.json.get("description")
-    assert "does not exist" in desc
-    assert client.response.status_code == 400
-
-
-def test_refuses_update_for_instance_with_operation(client, dns, service_instance):
-    dns.add_cname("_acme-challenge.bar.com")
-    dns.add_cname("_acme-challenge.foo.com")
-    factories.OperationFactory.create(service_instance=service_instance)
-    client.update_alb_instance("4321", params={"domains": "bar.com,foo.com"})
-
-    desc = client.response.json.get("description")
-
-    assert "in progress" in desc
-    assert client.response.status_code == 400
-
-
 def subtest_update_happy_path(
     client, dns, tasks, route53, iam_govcloud, simple_regex, alb
 ):
