+++ tests/integration/test_cdn_update_common_checks.py
@@ -86,203 +86,6 @@ def service_instance(service_instance_factory):
     "instance_model",
     [CDNServiceInstance, CDNDedicatedWAFServiceInstance],
 )
-def test_refuses_to_update_synchronously(instance_model, client):
-    client.update_instance(instance_model, "4321", accepts_incomplete="false")
-
-    assert "AsyncRequired" in client.response.body
-    assert client.response.status_code == 422
-
-
-@pytest.mark.parametrize(
-    "instance_model",
-    [CDNServiceInstance, CDNDedicatedWAFServiceInstance],
-)
-def test_refuses_to_update_synchronously_by_default(instance_model, client):
-    client.update_instance(instance_model, "4321", accepts_incomplete="")
-
-    assert "AsyncRequired" in client.response.body
-    assert client.response.status_code == 422
-
-
-@pytest.mark.parametrize(
-    "instance_model",
-    [CDNServiceInstance, CDNDedicatedWAFServiceInstance],
-)
-def test_refuses_to_update_with_duplicate_domains(
-    instance_model, client, dns, service_instance_factory, service_instance
-):
-    service_instance_factory.create(domain_names=["foo.com", "bar.com"])
-    dns.add_cname("_acme-challenge.example.com")
-    dns.add_cname("_acme-challenge.foo.com")
-
-    client.update_instance(
-        instance_model, "4321", params={"domains": "example.com, foo.com"}
-    )
-
-    assert "already exists" in client.response.body, client.response.body
-    assert client.response.status_code == 400, client.response.body
-
-
-@pytest.mark.parametrize(
-    "instance_model",
-    [CDNServiceInstance, CDNDedicatedWAFServiceInstance],
-)
-def test_duplicate_domain_check_ignores_self(
-    instance_model, client, dns, service_instance
-):
-    dns.add_cname("_acme-challenge.example.com")
-    dns.add_cname("_acme-challenge.foo.com")
-
-    client.update_instance(
-        instance_model, "4321", params={"domains": "example.com, foo.com"}
-    )
-
-    assert client.response.status_code == 202, client.response.body
-
-
-@pytest.mark.parametrize(
-    "instance_model",
-    [CDNServiceInstance, CDNDedicatedWAFServiceInstance],
-)
-def test_duplicate_domain_check_ignores_deactivated(
-    instance_model, client, dns, service_instance
-):
-    factories.CDNServiceInstanceFactory.create(
-        domain_names="foo.com", deactivated_at=datetime.utcnow()
-    )
-    dns.add_cname("_acme-challenge.foo.com")
-
-    client.update_instance(instance_model, "4321", params={"domains": "foo.com"})
-
-    assert client.response.status_code == 202, client.response.body
-
-
-@pytest.mark.parametrize(
-    "instance_model",
-    [CDNServiceInstance, CDNDedicatedWAFServiceInstance],
-)
-def test_refuses_to_update_without_any_acme_challenge_CNAMEs(
-    instance_model, client, service_instance
-):
-    client.update_instance(
-        instance_model, "4321", params={"domains": "bar.com,foo.com"}
-    )
-
-    desc = client.response.json.get("description")
-    assert "CNAME" in desc
-    assert "_acme-challenge.foo.com " in desc
-    assert "_acme-challenge.foo.com.domains.cloud.test" in desc
-    assert "_acme-challenge.bar.com " in desc
-    assert "_acme-challenge.bar.com.domains.cloud.test" in desc
-    assert client.response.status_code == 400
-
-
-@pytest.mark.parametrize(
-    "instance_model",
-    [CDNServiceInstance, CDNDedicatedWAFServiceInstance],
-)
-def test_refuses_to_update_without_one_acme_challenge_CNAME(
-    instance_model, client, dns, service_instance
-):
-    dns.add_cname("_acme-challenge.foo.com")
-    client.update_instance(
-        instance_model, "4321", params={"domains": "bar.com,foo.com"}
-    )
-
-    desc = client.response.json.get("description")
-    assert "CNAME" in desc
-    assert "_acme-challenge.foo.com" not in desc
-    assert "_acme-challenge.bar.com " in desc
-    assert "_acme-challenge.bar.com.domains.cloud.test" in desc
-    assert client.response.status_code == 400
-
-
-@pytest.mark.parametrize(
-    "instance_model",
-    [CDNServiceInstance, CDNDedicatedWAFServiceInstance],
-)
-def test_refuses_to_update_with_incorrect_acme_challenge_CNAME(
-    instance_model, client, dns, service_instance
-):
-    dns.add_cname("_acme-challenge.bar.com")
-    dns.add_cname("_acme-challenge.foo.com", target="INCORRECT")
-    client.update_instance(
-        instance_model, "4321", params={"domains": "bar.com,foo.com"}
-    )
-
-    desc = client.response.json.get("description")
-
-    assert "is set incorrectly" in desc
-
-    assert " _acme-challenge.foo.com " in desc
-    assert " _acme-challenge.foo.com.domains.cloud.test" in desc
-    assert " INCORRECT" in desc
-
-    assert " _acme-challenge.bar.com" not in desc
-    assert client.response.status_code == 400
-
-
-@pytest.mark.parametrize(
-    "instance_model",
-    [CDNServiceInstance, CDNDedicatedWAFServiceInstance],
-)
-def test_refuses_update_for_canceled_instance(
-    instance_model, client, dns, service_instance
-):
-    dns.add_cname("_acme-challenge.bar.com")
-    dns.add_cname("_acme-challenge.foo.com")
-    service_instance.deactivated_at = datetime.now()
-    db.session.add(service_instance)
-    db.session.commit()
-
-    client.update_instance(
-        instance_model, "4321", params={"domains": "bar.com,foo.com"}
-    )
-
-    desc = client.response.json.get("description")
-    assert "canceled" in desc
-    assert client.response.status_code == 400
-
-
-@pytest.mark.parametrize(
-    "instance_model",
-    [CDNServiceInstance, CDNDedicatedWAFServiceInstance],
-)
-def test_refuses_update_for_nonexistent_instance(instance_model, client, dns):
-    dns.add_cname("_acme-challenge.bar.com")
-    dns.add_cname("_acme-challenge.foo.com")
-    client.update_instance(
-        instance_model, "4321", params={"domains": "bar.com,foo.com"}
-    )
-    desc = client.response.json.get("description")
-    assert "does not exist" in desc
-    assert client.response.status_code == 400
-
-
-@pytest.mark.parametrize(
-    "instance_model",
-    [CDNServiceInstance, CDNDedicatedWAFServiceInstance],
-)
-def test_refuses_update_for_instance_with_operation(
-    instance_model, client, dns, service_instance
-):
-    dns.add_cname("_acme-challenge.bar.com")
-    dns.add_cname("_acme-challenge.foo.com")
-    factories.OperationFactory.create(service_instance=service_instance)
-    client.update_instance(
-        instance_model, "4321", params={"domains": "bar.com,foo.com"}
-    )
-
-    desc = client.response.json.get("description")
-
-    assert "in progress" in desc
-    assert client.response.status_code == 400
-
-
-@pytest.mark.parametrize(
-    "instance_model",
-    [CDNServiceInstance, CDNDedicatedWAFServiceInstance],
-)
 def test_provision_sets_default_origin_if_provided_as_none(
     instance_model, client, service_instance
 ):
